fix: deadlock when recording snapshots in editor
Waiting for disposal task that didn't use ConfigureAwait(false) on the main thread caused a deadlock. This fix is based on removing unnecessary uses of async/await that wrapped code that was supposed to be synchronous anyway. Moving synchronous waiting to the lowest possible level removed the need to add multiple calls to ConfigureAwait and better exposed where the responsibility of ensuring synchronous execution free of deadlock actually lies.